Create Product constuctor, and split product from page catalog to page Product, and some refactoring
