+++ CatalogModel/InMemoryCatalog.go
@@ -24,7 +24,7 @@ func (InMemoryCatalogFactory) CreateCatalog() Catalog {
 }
 
 //
-func (catalog InMemoryCatalog) AddNewProduct(product *Product) (int, error) {
+func (catalog *InMemoryCatalog) AddNewProduct(product *Product) (int, error) {
 	a := len(catalog.products)
 	product.id = a + 1
 	catalog.products[product.id] = product
@@ -32,7 +32,7 @@ func (catalog InMemoryCatalog) AddNewProduct(product *Product) (int, error) {
 	return 0, errors.New("cannot add product")
 }
 
-func (catalog InMemoryCatalog) DeleteProductById(cameId int) error {
+func (catalog *InMemoryCatalog) DeleteProductById(cameId int) error {
 	for key := cameId; key < len(catalog.products); key++ {
 		catalog.products[key] = catalog.products[key+1]
 		catalog.products[key].id--
@@ -42,11 +42,11 @@ func (catalog InMemoryCatalog) DeleteProductById(cameId int) error {
 	return errors.New("can't edit product")
 }
 
-func (catalog InMemoryCatalog) GetAll() map[int]*Product {
+func (catalog *InMemoryCatalog) GetAll() map[int]*Product {
 	return catalog.products
 }
 
-func (catalog InMemoryCatalog) EditProduct(cameId int, name string, count int64, price float64) (int, error) {
+func (catalog *InMemoryCatalog) EditProduct(cameId int, name string, count int64, price float64) (int, error) {
 
 	catalog.products[cameId].id = cameId
 	catalog.products[cameId].name = name
@@ -57,6 +57,6 @@ func (catalog InMemoryCatalog) EditProduct(cameId int, name string, count int64,
 	return cameId, errors.New("can't edit product")
 }
 
-func (catalog InMemoryCatalog) GetProductByID(cameId int) (*Product, error) {
+func (catalog *InMemoryCatalog) GetProductByID(cameId int) (*Product, error) {
 	return catalog.products[cameId], errors.New("product not found")
 }
